Final Product (rev)

+++ LAB/delete.php
@@ -1,6 +1,6 @@
 <?php
 if (isset($_POST['delete'])){
-    $delete = "DELETE FROM todolist WHERE task = (SELECT task FROM todolist LIMIT 1 OFFSET " . $_POST['id'] . ")";
+    $delete = "DELETE FROM todolist WHERE task = (SELECT task FROM todolist WHERE user_id = '" . $_COOKIE['user_id']. "'LIMIT 1 OFFSET " . $_POST['id'] . ")";
     $result = $db->query($delete);
 }
 ?>
